% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{ustyc}
\alias{ustyc}
\alias{ustyc-package}
\title{US Treasury yield curve.}
\description{
Fetches US Treasury yield curve data and transforms to data frame.
}
\details{
Downloads the US Treasury yield curve data in its original XML format,
then translates that data into a data frame with dates as row names and maturities
as column names.
}
\author{
Matt Barry \email{mrb@softisms.com}
}
\references{
US Treasury resource web site \url{http://www.treasury.gov/resource-center/data-chart-center/interest-rates/Pages/TextView.aspx?data=yield}
}
\keyword{curve}
\keyword{treasury}
\keyword{yield}

